+++ uni_modules/uview-ultra/components/up-swiper/up-swiper.vue
@@ -27,6 +27,7 @@
 			:autoplay="autoplay"
 			:current="current"
 			:currentItemId="currentItemId"
+			:vertical="vertical"
 			:previousMargin="addUnit(previousMargin)"
 			:nextMargin="addUnit(nextMargin)"
 			:acceleration="acceleration"